refactor(frontend): move useProposedChange to ui/hooks/
Hooks belong in ui/hooks/ per the entity-structure guideline; the atom
stays in stores/.
