+++ frontend/app/src/entities/proposed-changes/ui/action-button/pc-action-button.tsx
@@ -7,13 +7,13 @@ import { inputStyle } from "@/shared/components/ui/style";
 import { classNames } from "@/shared/utils/common";
 
 import { useAuth } from "@/entities/authentication/ui/useAuth";
-import { useProposedChange } from "@/entities/proposed-changes/stores/proposedChanges.atom";
 import { ActionComboboxList } from "@/entities/proposed-changes/ui/action-button/pc-action-combobox-list";
 import { CloseButton } from "@/entities/proposed-changes/ui/action-button/pc-close-button";
 import { DraftButton } from "@/entities/proposed-changes/ui/action-button/pc-draft-button";
 import { MergeButton } from "@/entities/proposed-changes/ui/action-button/pc-merge-button";
 import { PcPlaceholderButton } from "@/entities/proposed-changes/ui/action-button/pc-placeholder-button";
 import type { ProposedChangeActionButtonProps } from "@/entities/proposed-changes/ui/action-button/types";
+import { useProposedChange } from "@/entities/proposed-changes/ui/hooks/use-proposed-change";
 import { PcActionsContext } from "@/entities/proposed-changes/ui/pc-actions-permissions-context";
 import { useGetProposedChangeAvailableActions } from "@/entities/proposed-changes/ui/queries/get-proposed-change-available-actions.query";
