+++ frontend/app/src/entities/proposed-changes/ui/action-button/pc-merge-button.tsx
@@ -11,7 +11,7 @@ import { useNavigateAfterBranchRemoval } from "@/entities/branches/ui/hooks/use-
 import { useConfig } from "@/entities/config/ui/config-provider";
 import { useUpdateObjectMutation } from "@/entities/nodes/object/ui/queries/update-object.mutation";
 import { MERGE_STATE } from "@/entities/proposed-changes/constants";
-import { useProposedChange } from "@/entities/proposed-changes/stores/proposedChanges.atom";
+import { useProposedChange } from "@/entities/proposed-changes/ui/hooks/use-proposed-change";
 import { usePcActionsContext } from "@/entities/proposed-changes/ui/pc-actions-permissions-context";
 
 import type { ProposedChangeActionButtonProps } from "./types";
